Chaos of War Scenario update Break station placement out into a separate file Some configuration options on the up front screen Include explanations of configuration options via buttons on the GM screen Add advance intel option
